requirements.txt
mysql-connector-python
python-dotenv
bs4
